+++ tests/views/test_assignments_view.py
@@ -34,6 +34,20 @@ class TestAssignmentView(BaseVersionedTestView):
         'payout_currency': 'USD'
     }
 
+    def test_put_invalid_asset_tyoe(self, app, obj_payload):
+        """Asset type should match one of the possible values."""
+        payload = obj_payload
+        obj_id = payload['id']
+        payload['asset_types'] = ['Foobar']
+        request = app.put_json('{base}/{id}'.format(base=self.base_path, id=obj_id),
+                               payload, headers=self.headers, status=400)
+        result = request.json
+        error = result['errors'][0]
+        assert result['status'] == 'error'
+        assert error['name'] == 'asset_types'
+        assert error['location'] == 'body'
+        assert 'Invalid type of asset' in error['description']
+
     def test_workflow(self, app, session, instance_obj):
         """Test workflow endpoints."""
         payload = {